Add a shared OpenAI trait and move shared code into there so we aren't duplicating that. Change our auth callback to use the cheapest model, since API keys work the same for all endpoints
